Reorganise master term models, update torsion model.
