+++ src/gui/models/masterImproperModel.h
@@ -1,0 +1,33 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2023 Team Dissolve and contributors
+
+#pragma once
+
+#include "classes/speciesimproper.h"
+#include "gui/models/masterTermModel.h"
+#include "templates/optionalref.h"
+
+// MasterImproper model
+class MasterImproperModel : public MasterTermModel
+{
+    Q_OBJECT
+
+    public:
+    MasterImproperModel(QObject *parent = nullptr);
+
+    private:
+    // Source term data
+    OptionalReferenceWrapper<std::vector<std::shared_ptr<MasterImproper>>> sourceData_;
+
+    public:
+    // Set source data
+    void setSourceData(std::vector<std::shared_ptr<MasterImproper>> &bonds);
+
+    /*
+     * QAbstractItemModel overrides
+     */
+    public:
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
+    QVariant getTermData(int row, MasterTermModelData::DataType dataType) const override;
+    bool setTermData(int row, MasterTermModelData::DataType dataType, const QVariant &value) override;
+};
