+++ src/gui/models/masterTorsionModel.h
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2023 Team Dissolve and contributors
+
+#pragma once
+
+#include "classes/speciestorsion.h"
+#include "gui/models/masterTermModel.h"
+#include "templates/optionalref.h"
+
+// MasterTorsion model
+class MasterTorsionModel : public MasterTermModel
+{
+    Q_OBJECT
+
+    public:
+    MasterTorsionModel(QObject *parent = nullptr);
+
+    private:
+    // Source term data
+    OptionalReferenceWrapper<std::vector<std::shared_ptr<MasterTorsion>>> sourceData_;
+
+    public:
+    // Set source data
+    void setSourceData(std::vector<std::shared_ptr<MasterTorsion>> &terms);
+
+    /*
+     * QAbstractItemModel overrides
+     */
+    public:
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
+    int columnCount(const QModelIndex &parent = QModelIndex()) const override;
+    QVariant getTermData(int row, MasterTermModelData::DataType dataType) const override;
+    bool setTermData(int row, MasterTermModelData::DataType dataType, const QVariant &value) override;
+};
